+++ modules/apps/discord.nix
@@ -50,6 +50,7 @@
           expressionCloner.enable = true;
           favoriteEmojiFirst.enable = true;
           favoriteGifSearch.enable = true;
+          fakeNitro.enable = true;
           fixImagesQuality.enable = true;
           fixSpotifyEmbeds.enable = true;
           friendsSince.enable = true;