+++ modules/desktop/hyprland/_features/files.nix
@@ -1,3 +1,4 @@
+# File Manager
 {
   util ? null,
   files ? null,