+++ modules/desktop/hyprland/_features/login.nix
@@ -1,3 +1,4 @@
+# Login Greeter
 _: {
   nixos = {
     config,