+++ modules/desktop/hyprland/_features/main.nix
@@ -1,3 +1,4 @@
+# Compositor
 {
   util ? null,
   files ? null,