+++ modules/desktop/hyprland/_features/media.nix
@@ -1,3 +1,4 @@
+# Media
 {util ? null, ...}: {
   nixos = {pkgs, ...}: {
     environment.systemPackages = with pkgs; [